secure routes requiring authorization

+++ server-side/routes/api.controller.js
@@ -13,46 +13,61 @@ var getPictures = (req, res) => {
 };
 
 var addPicture = (req, res) => {
-    var picture = new Picture(req.body);
-    picture.save();
-    res.status(201).json(picture);
+    if (req.user) {
+        var picture = new Picture(req.body);
+        picture.save();
+        res.status(201).json(picture);
+    }
+    else {
+        res.sendStatus(401);
+    }
 };
 
 var updatePicture = (req, res) => {
-    var id = req.params.id;
-    var updatedProperties = req.body;
-    Picture.findById(id, (err, picture) => {
-        if (err) {
-            res.sendStatus(500);
-        }
-        else if (!picture) {
-            res.sendStatus(404);
-        }
-        else {
-            var updated = _.assign(picture, updatedProperties);
-        }
-        updated.save(err => {
+    if (req.user) {
+        var id = req.params.id;
+        var updatedProperties = req.body;
+        Picture.findById(id, (err, picture) => {
             if (err) {
                 res.sendStatus(500);
             }
-            else {
-                res.status(200).json(updated);
+            else if (!picture) {
+                res.sendStatus(404);
             }
+            else {
+                var updated = _.assign(picture, updatedProperties);
+            }
+            updated.save(err => {
+                if (err) {
+                    res.sendStatus(500);
+                }
+                else {
+                    res.status(200).json(updated);
+                }
+            });
         });
-    });
+    }
+    else {
+        res.sendStatus(401);
+    }
 };
 
 var removePicture = (req, res) => {
-    var id = req.params.id;
-    Picture.findByIdAndRemove(id, (err, picture) => {
-        if (err) {
-            console.log(err);
-            res.sendStatus(500);
-        }
-        else {
-            res.status(200).json(picture);
-        }
-    });
+    if (req.user) {
+        var id = req.params.id;
+        Picture.findByIdAndRemove(id, (err, picture) => {
+            if (err) {
+                console.log(err);
+                res.sendStatus(500);
+            }
+            else {
+                res.status(200).json(picture);
+            }
+        });
+    }
+    else {
+        res.sendStatus(401);
+    }
 };
 
 var apiController = {
